move out ai api request from ai folder to make pr diff more clear
